pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "production-rag-system"
version = "0.1.0"
description = "Production-grade RAG system with hybrid retrieval, citation enforcement, and CI-gated quality"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Abhishek Rastogi" },
]

dependencies = [
    "anthropic>=0.80.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",
    "tiktoken>=0.7.0",
    "pypdf>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "structlog>=24.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]
# Phase 2 additions
hybrid = [
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
]

[project.scripts]
rag-ingest = "rag.cli:ingest"
rag-query = "rag.cli:query"

[tool.hatch.build.targets.wheel]
packages = ["src/rag"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short --cov=rag --cov-report=term-missing"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
